requirements.txt
gradio
openai
python-dotenv
colorama
pyfiglet